requirements.txt
celery
redis
beautifulsoup4
requests
psycopg2
python-dotenv
